// miniproject1-frontend/types/api.ts
export interface components {
    schemas: {
// ...
        PharmacyDetailResponse: {
            /** Format: int64 */
            id?: number;
            name?: string;
            addressRoad?: string;
            addressJibun?: string;
            /** Format: double */
            lat?: number;
            /** Format: double */
            lng?: number;
            phone?: string;
            businessHours?: {
                [key: string]: string[] | null;
            };
            /** Format: int64 */
            distanceM?: number;
            region?: components["schemas"]["RegionInfo"];
            drugPrices?: components["schemas"]["DrugPriceItem"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}